GHI-544 refactor ReviewTransactionViewController
